Rename rest of the cluster dir
