+++ src/app/cluster/cluster-details/revoke-token/style.scss
