+++ src/app/cluster/cluster-details/revoke-token/template.html
