+++ src/app/cluster/cluster-details/share-kubeconfig/component.spec.ts
