+++ src/app/cluster/cluster-details/share-kubeconfig/style.scss
