+++ src/app/cluster/cluster-details/share-kubeconfig/template.html
